Finished task ownership section

+++ src/tasks/tasks.service.ts
@@ -17,9 +17,9 @@ export class TasksService {
         private taskRepository:TaskRepository, 
     ) { }
 
-      async getTaskById(id: number): Promise<Task> {
+      async getTaskById(id: number, user:User): Promise<Task> {
 
-          const foundTask = await this.taskRepository.findOne(id);
+          const foundTask = await this.taskRepository.findOne({where:{id, userId:user.id}});
 
           if (!foundTask) {
               throw new NotFoundException(`task with id ${id} not found`);
@@ -28,7 +28,7 @@ export class TasksService {
 
       }
     
-      async deleteTask(id: number){
+      async deleteTask(id: number, user:User){
 
         // const foundTask = await this.taskRepository.findOne(id);
 
@@ -36,8 +36,8 @@ export class TasksService {
         //     throw new NotFoundException(`task with id ${id} not found`);
         // }
         //   foundTask.delete();
-          const result = await this.taskRepository.delete(id)
-          console.log(result)
+        const result = await this.taskRepository.delete({ id, userId: user.id })
+        
           if (result.affected === 0) {
             throw new NotFoundException(`task with id ${id} not found`);
               
@@ -62,30 +62,24 @@ export class TasksService {
     };
 
   
-    // deleteTask(id: string) {
-        
-    //     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    //     const foundTask = this.getTaskById(id);
-    //     this.tasks = this.tasks.filter(task=> task.id !==id)
-    // }
 
-    async updateTask(id: number, status: TaskStatus) :Promise<Task>{
+    async updateTask(id: number, status: TaskStatus, user:User) :Promise<Task>{
 
-        const task =  await this.getTaskById(id);
+        const task =  await this.getTaskById(id,user);
         task.status = status;
         await task.save();
         return task;
 
     }
 
-    async searchTask(filterDto: getTasksFilterDto): Promise<Task[]>{
+    async searchTask(filterDto: getTasksFilterDto, user: User): Promise<Task[]>{
 
         // const filtered: Task[] = status? this.tasks.filter(task => task.status === status
         //     && (task.title.includes(search) || task.description.includes(search))) :
         //     this.tasks.filter(task=>task.title.includes(search) || task.description.includes(search))
     
         // return filtered
-        const res = await this.taskRepository.searchTask(filterDto)
+        const res = await this.taskRepository.searchTask(filterDto, user)
         return res
     }
